リポジトリ選択をコンポーネントとして切り出した
